refactor: move physics helpers to module
